Refactor Shopily dashboard view into modular helpers
