+++ src/ui/views/browser/components/shopily/views/dashboard/nicheFormatting.js
@@ -1,0 +1,7 @@
+export function formatNicheDelta(delta, formatPercent = value => String(value ?? '')) {
+  if (delta === null || delta === undefined) return '';
+  const numeric = Number(delta);
+  if (!Number.isFinite(numeric) || numeric === 0) return '';
+  const icon = numeric > 0 ? '⬆️' : '⬇️';
+  return `${icon} ${Math.abs(Math.round(numeric * 100))}%`;
+}
